requirements.txt
asgiref==3.8.1
Django==4.2.21
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
